Adjust path move ./res to ./src/res script-build-output: ./src -> ./gh-pages

+++ scripts/build.js
@@ -16,22 +16,22 @@ const _outputDir = path.resolve(cwd, process.argv[3]);
 const hamlRegex = /\.haml$/i;
 
 (function recursiveCompile(inputDir, outputDir) {
+  if (fs.existsSync(outputDir)) {
+    if (!fs.statSync(outputDir).isDirectory()) {
+      throw `Path '${outputDir}' should be a folder.`;
+    }
+  } else {
+    fs.mkdirSync(outputDir);
+  }
   const names = fs.readdirSync(inputDir);
   names.forEach(name => {
     const subInputPath = path.join(inputDir, name);
     const subOutputPath = path.join(outputDir, name);
     const stat = fs.statSync(subInputPath);
     if (stat.isDirectory()) {
-      if (fs.existsSync(subOutputPath)) {
-        if (!fs.statSync(subOutputPath).isDirectory()) {
-          throw `Path '${subOutputPath}' should be a folder.`;
-        }
-      } else {
-        fs.mkdirSync(subOutputPath);
-      }
       recursiveCompile(subInputPath, subOutputPath);
-    } else if (stat.isFile()) {
-      if (hamlRegex.test(name)) {
+    } else {
+      if (stat.isFile() && hamlRegex.test(name)) {
         const content = fs.readFileSync(subInputPath, 'utf8');
         fs.writeFileSync(path.join(outputDir, name.replace(hamlRegex, '.html')), render(content));
       } else {
